add infantry moving towards the player main base

+++ src/components/Movement.hpp
@@ -5,10 +5,9 @@
 
 struct Movement
 {
-    Vec2i direction;
+    Vec2d direction;
     double speed;
     bool move;
 };
-
 
 #endif // MOVEMENT_HPP
